ProcessCommand: cleanup
