+++ src/FileSystem/PhpFilesFinder.php
@@ -2,6 +2,7 @@
 
 namespace Rector\FileSystem;
 
+use Rector\Exception\FileSystem\DirectoryNotFoundException;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 
@@ -13,6 +14,8 @@ final class PhpFilesFinder
      */
     public function findInDirectories(array $directories): array
     {
+        $this->ensureDirectoriesExist($directories);
+
         $finder = Finder::create()
             ->files()
             ->name('*.php')
@@ -23,5 +26,22 @@ final class PhpFilesFinder
             ->in($directories);
 
         return iterator_to_array($finder->getIterator());
+    }
+
+    /**
+     * @param string[] $directories
+     */
+    private function ensureDirectoriesExist(array $directories): void
+    {
+        foreach ($directories as $directory) {
+            if (file_exists($directory)) {
+                continue;
+            }
+
+            throw new DirectoryNotFoundException(sprintf(
+                'Directory "%s" was not found.',
+                $directory
+            ));
+        }
     }
 }
